Clean up build script, update default test suite
